// server/enums/applicationRole.ts
enum ApplicationRole {
  ROLE_LEARNING_AND_WORK_PROGRESS_MANAGER = 'ROLE_LEARNING_AND_WORK_PROGRESS_MANAGER',
  ROLE_LEARNING_AND_WORK_PROGRESS_CONTRIBUTOR = 'ROLE_LEARNING_AND_WORK_PROGRESS_CONTRIBUTOR',
  // ROLE_EDUCATION_WORK_PLAN_EDITOR is a deprecated role
  ROLE_EDUCATION_WORK_PLAN_EDITOR = 'ROLE_EDUCATION_WORK_PLAN_EDITOR',
}

export default ApplicationRole