// server/enums/applicationAction.ts
enum ApplicationAction {
  RECORD_INDUCTION = 'RECORD_INDUCTION',
  EXEMPT_INDUCTION = 'EXEMPT_INDUCTION',
  REMOVE_INDUCTION_EXEMPTION = 'REMOVE_INDUCTION_EXEMPTION',
  UPDATE_INDUCTION = 'UPDATE_INDUCTION',
  RECORD_REVIEW = 'RECORD_REVIEW',
  EXEMPT_REVIEW = 'EXEMPT_REVIEW',
  REMOVE_REVIEW_EXEMPTION = 'REMOVE_REVIEW_EXEMPTION',
  CREATE_GOALS = 'CREATE_GOALS',
  UPDATE_GOALS = 'UPDATE_GOALS',
}

export default ApplicationAction